Make inventory code a separate component
